ipcams/requirements.txt
anyio==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
h11==0.12.0
hikvisionapi==0.3.2
httpcore==0.14.7
httpx==0.22.0
idna==3.3
numpy==1.21.5
opencv-python==4.5.5.62
python-dotenv==0.19.2
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
typing_extensions==4.1.1
urllib3==1.26.8
xmltodict==0.12.0
